+++ src/commands/register.rs
@@ -2,6 +2,7 @@ use crate::models::{conversion, HasName, Ledger, Posting, PostingType};
 use crate::models::{Balance, Money};
 use crate::parser::value_expr::build_root_node_from_expression;
 use crate::{filter, CommonOpts};
+use chrono::{DateTime, NaiveDate, Utc};
 use colored::Colorize;
 use std::collections::HashMap;
 use std::convert::TryFrom;
@@ -22,7 +23,7 @@ pub fn execute(
     };
 
     let mut balance = Balance::new();
-
+    let today = Utc::now().naive_local().date();
     let size = terminal_size();
     let mut width: usize = 80;
     if let Some((Width(w), _)) = size {
@@ -139,17 +140,22 @@ pub fn execute(
         for p in postings_vec.iter() {
             counter += 1;
             if counter == 1 {
+                let mut date_str = format!("{}", t.date.unwrap().format(&options.date_format)).normal();
+                if t.date.unwrap() > today {
+                    date_str = date_str.green();
+                }
+
                 match t.get_payee(&ledger.payees) {
                     Some(payee) => print!(
                         "{:w1$}{:width$}",
-                        format!("{}", t.date.unwrap().format(&options.date_format)).green(),
+                        date_str,
                         clip(&format!("{} ", payee), w_description),
                         width = w_description,
                         w1 = w_date
                     ),
                     None => print!(
                         "{:w1$}{:width$}",
-                        format!("{}", t.date.unwrap().format(&options.date_format)).green(),
+                        date_str,
                         clip(&format!("{} ", ""), w_description),
                         width = w_description,
                         w1 = w_date