+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
-## [0.33.0] - xxx
+## [0.33.0] - 2021-09-26
 ## Changed
 - Nicer error messages (without Rust trace) when there is a missing file.
 ## Fixed